+++ src/Control/Eff/Exception.hs
@@ -22,7 +22,7 @@ module Control.Eff.Exception ( Exc (..)
 
 import Control.Eff
 import Control.Eff.Lift
-import Data.OpenUnion51
+import Data.OpenUnion
 
 import Control.Monad (void)
