+++ src/Control/Eff/Fresh.hs
@@ -14,7 +14,7 @@ module Control.Eff.Fresh( Fresh (..)
                         ) where
 
 import Control.Eff
-import Data.OpenUnion51
+import Data.OpenUnion
 
 -- There are three possible implementations
 -- The first one uses State Fresh where
