+++ src/Control/Eff/Writer/Strict.hs
@@ -16,7 +16,7 @@ module Control.Eff.Writer.Strict ( Writer(..)
                                ) where
 
 import Control.Eff
-import Data.OpenUnion51
+import Data.OpenUnion
 import Data.FTCQueue1
 
 import Data.Monoid
