+++ src/Data/OpenUnion.hs
@@ -1,104 +1,139 @@
 {-# OPTIONS_HADDOCK show-extensions #-}
 
-{-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE TypeOperators #-}
-{-# LANGUAGE CPP #-}
-{-# LANGUAGE Safe #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE DataKinds, PolyKinds #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE MultiParamTypeClasses, FlexibleInstances, FlexibleContexts #-}
+{-# LANGUAGE OverlappingInstances #-}
+{-# LANGUAGE Trustworthy #-}
+{-# OPTIONS_GHC -fno-warn-warnings-deprecations -Wwarn #-}
 
--- Only for SetMember below, when emulating Monad Transformers
+-- Only for MemberU below, when emulating Monad Transformers
 {-# LANGUAGE FunctionalDependencies, UndecidableInstances #-}
-{-# LANGUAGE UndecidableSuperClasses #-}
-{-# LANGUAGE FlexibleContexts #-}
 
-#if __GLASGOW_HASKELL__ >= 708
-#define Typeable1 Typeable
-#endif
--- | Original work at <http://okmij.org/ftp/Haskell/extensible/OpenUnion1.hs>
--- and <http://okmij.org/ftp/Haskell/extensible/OpenUnion2.hs>.
--- Open unions (type-indexed co-products) for extensible effects.
+-- | Open unions (type-indexed co-products) for extensible effects
+-- All operations are constant-time, and there is no Typeable constraint
 --
--- TODO: see if we can do away with Typeable constraints, perhaps by
--- incorporating ideas from <http://okmij.org/ftp/Haskell/extensible/TList.hs>
-module Data.OpenUnion(
-  -- * Classes
-  -- | @`Member` t r@ specifies whether @t@ is present anywhere in the sum type
-  -- @r@, where @t@ is some effectful type, e.g. @`Lift` `IO`@, @`State` Int`@
-  Member
-  -- ** Monad transformer related
-  -- | @`SetMember` set t r@ is used to emulate monad transformers.
-  , SetMember
-    -- * Type-indexed co-product
-    -- ** Datatypes
-  , Union
-  , (:>)
-    -- ** Functions
-  , inj
-  , prj
-  , prjForce
-  , decomp
-  , unsafeReUnion
-  , weaken
-  ) where
-
-import Data.OpenUnion.Imports
-import Data.Typeable
-
-infixl 4 <?>
-
--- | infix form of `fromMaybe`.
-(<?>) :: Maybe a -> a -> a
-Just a <?> _ = a
-_ <?> a = a
-
--- | for the sake of @gcast1@ used below in @`prj`@
-newtype Id a = Id { runId :: a }
-  deriving Typeable
-
--- | `SetMember` is similar to `Member`, but it allows types to belong to a
--- \"set\". For every set, only one member can be in @r@ at any given time.
--- This allows us to specify exclusivity and uniqueness among arbitrary effects:
+-- This is a variation of OpenUion5.hs, which relies on overlapping
+-- instances instead of closed type families. Closed type families
+-- have their problems: overlapping instances can resolve even
+-- for unground types, but closed type families are subject to a
+-- strict apartness condition.
 --
--- > -- Terminal effects (effects which must be run last)
--- > data Terminal
--- >
--- > -- Make Lifts part of the Terminal effects set.
--- > -- The fundep assures that there can only be one Terminal effect for any r.
--- > instance Member (Lift m) r => SetMember Terminal (Lift m) r
--- >
--- > -- Only allow a single unique Lift effect, by making a "Lift" set.
--- > instance Member (Lift m) r => SetMember Lift (Lift m) r
-class (Member t r) => SetMember set (t :: * -> *) r | r set -> t
-instance (MemberU set t r) => SetMember set t r
+-- This implementation is very similar to OpenUnion1.hs, but without
+-- the annoying Typeable constraint. We sort of emulate it:
+--
+-- Our list r of open union components is a small Universe.
+-- Therefore, we can use the Typeable-like evidence in that
+-- universe. We hence can define
+--
+-- @
+-- data Union r v where
+--   Union :: t v -> TRep t r -> Union r v -- t is existential
+-- @
+-- where
+--
+-- @
+-- data TRep t r where
+--   T0 :: TRep t (t ': r)
+--   TS :: TRep t r -> TRep (any ': r)
+-- @
+-- Then Member is a type class that produces TRep
+-- Taken literally it doesn't seem much better than
+-- OpenUinion41.hs. However, we can cheat and use the index of the
+-- type t in the list r as the TRep. (We will need UnsafeCoerce then).
 
-{-# INLINE inj #-}
--- | Construct a Union.
-inj :: (Functor t, Typeable1 t, Member t r) => t v -> Union r v
-inj = Union
+-- The interface is the same as of other OpenUnion*.hs
+module Data.OpenUnion (Union, inj, prj, decomp,
+                   Member, MemberU2, weaken
+                  ) where
 
-{-# INLINE prj #-}
--- | Try extracting the contents of a Union as a given type.
-prj :: (Typeable1 t, Member t r) => Union r v -> Maybe (t v)
-prj (Union v) = runId `fmap` gcast1 (Id v)
+import Unsafe.Coerce(unsafeCoerce)
 
-{-# INLINE prjForce #-}
--- | Extract the contents of a Union as a given type.
--- If the Union isn't of that type, a runtime error occurs.
-prjForce :: (Typeable1 t, Member t r) => Union r v -> (t v -> a) -> a
-prjForce u f = f `fmap` prj u <?> error "prjForce with an invalid type"
+-- The data constructors of Union are not exported
 
-{-# INLINE decomp #-}
--- | Try extracting the contents of a Union as a given type.
--- If we can't, return a reduced Union that excludes the type we just checked.
-decomp :: Typeable1 t => Union (t :> r) v -> Either (Union r v) (t v)
-decomp u = Right `fmap` prj u <?> Left (unsafeReUnion u)
+-- Strong Sum (Existential with the evidence) is an open union
+-- t is can be a GADT and hence not necessarily a Functor.
+-- Int is the index of t in the list r; that is, the index of t in the
+-- universe r
+data Union (r :: [ * -> * ]) v where
+  Union :: {-# UNPACK #-} !Int -> t v -> Union r v
 
-{-# INLINE weaken #-}
-weaken :: (Typeable1 t, Functor t) => Union r w -> Union (t :> r) w
-weaken (Union x) = Union x
+{-# INLINE prj' #-}
+{-# INLINE inj' #-}
+inj' :: Int -> t v -> Union r v
+inj' = Union
 
-{-# INLINE unsafeReUnion #-}
--- | Juggle types for a Union. Use cautiously.
-unsafeReUnion :: Union r w -> Union t w
-unsafeReUnion (Union v) = Union v
+prj' :: Int -> Union r v -> Maybe (t v)
+prj' n (Union n' x) | n == n'   = Just (unsafeCoerce x)
+                    | otherwise = Nothing
+
+newtype P t r = P{unP :: Int}
+
+class (FindElem t r) => Member (t :: * -> *) r where
+  inj :: t v -> Union r v
+  prj :: Union r v -> Maybe (t v)
+
+{-
+-- Optimized specialized instance
+instance Member t '[t] where
+  {-# INLINE inj #-}
+  {-# INLINE prj #-}
+  inj x           = Union 0 x
+  prj (Union _ x) = Just (unsafeCoerce x)
+-}
+
+instance (FindElem t r) => Member t r where
+  {-# INLINE inj #-}
+  {-# INLINE prj #-}
+  inj = inj' (unP $ (elemNo :: P t r))
+  prj = prj' (unP $ (elemNo :: P t r))
+
+
+{-# INLINE [2] decomp #-}
+decomp :: Union (t ': r) v -> Either (Union r v) (t v)
+decomp (Union 0 v) = Right $ unsafeCoerce v
+decomp (Union n v) = Left  $ Union (n-1) v
+
+
+-- Specialized version
+{-# RULES "decomp/singleton"  decomp = decomp0 #-}
+{-# INLINE decomp0 #-}
+decomp0 :: Union '[t] v -> Either (Union '[] v) (t v)
+decomp0 (Union _ v) = Right $ unsafeCoerce v
+-- No other case is possible
+
+weaken :: Union r w -> Union (any ': r) w
+weaken (Union n v) = Union (n+1) v
+
+-- | Find an index of an element in a `list'
+-- The element must exist
+-- This is essentially a compile-time computation.
+-- Using overlapping instances here is OK since this class is private to this
+-- module
+class FindElem (t :: * -> *) r where
+  elemNo :: P t r
+
+instance FindElem t (t ': r) where
+  elemNo = P 0
+
+instance FindElem t r => FindElem t (t' ': r) where
+  elemNo = P $ 1 + (unP $ (elemNo :: P t r))
+
+
+-- | Using overlapping instances here is OK since this class is private to this
+-- module
+class EQU (a :: k) (b :: k) p | a b -> p
+instance EQU a a 'True
+instance (p ~ 'False) => EQU a b p
+
+-- | This class is used for emulating monad transformers
+class Member t r => MemberU2 (tag :: k -> * -> *) (t :: * -> *) r | tag r -> t
+instance (EQU t1 t2 p, MemberU' p tag t1 (t2 ': r)) => MemberU2 tag t1 (t2 ': r)
+
+class Member t r =>
+      MemberU' (f::Bool) (tag :: k -> * -> *) (t :: * -> *) r | tag r -> t
+instance MemberU' 'True tag (tag e) (tag e ': r)
+instance (Member t (t' ': r), MemberU2 tag t r) =>
+           MemberU' 'False tag t (t' ': r)
